move random functions to libmutt
