+++ mutt/random.h
@@ -1,0 +1,33 @@
+/**
+ * @file
+ * Random number/string functions
+ *
+ * @authors
+ * Copyright (C) 2020 Richard Russon <rich@flatcap.org>
+ *
+ * @copyright
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 2 of the License, or (at your option) any later
+ * version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef MUTT_LIB_RANDOM_H
+#define MUTT_LIB_RANDOM_H
+
+#include <stdint.h>
+#include <stddef.h>
+
+uint32_t mutt_rand32(void);
+uint64_t mutt_rand64(void);
+void     mutt_rand_base32(char *buf, size_t buflen);
+
+#endif /* MUTT_LIB_RANDOM_H */
